`Signature` struct  is not in the right place
Move from "transaction.rs" into "signature.rs".
