+++ src/core/mod.rs
@@ -7,6 +7,6 @@ mod transaction;
 
 pub use self::address::{ADDRESS_BYTES, Address};
 pub use self::error::Error;
-pub use self::signature::{ECDSA_SIGNATURE_BYTES, PRIVATE_KEY_BYTES, PrivateKey};
-pub use self::transaction::{Signature, Transaction};
+pub use self::signature::{ECDSA_SIGNATURE_BYTES, PRIVATE_KEY_BYTES, PrivateKey, Signature};
+pub use self::transaction::Transaction;
 use super::util;
